+++ bakedanuki/bakedanuki-util/python/bd_util/_test/test.py
@@ -1,5 +1,4 @@
 # coding:utf-8
-import random
 
 # self
 import bd_util as bdu
@@ -10,48 +9,71 @@ logger = u_logger.get_logger(__name__, level=u_logger.DEBUG)
 
 def main():
     nodes = bdu.Nodes()
-    src_p = nodes.create.transform(name="src_parent")
-    src, _ = nodes.create.with_transform.locator(name="src", parent=src_p)
+    src = nodes.create.transform(name="src")
+    dst = nodes.create.joint(name="dst")
 
-    dst_p = nodes.create.transform(name="dst_parent")
-    dst_a, _ = nodes.create.with_transform.locator(name="dst_a", parent=dst_p)
-    dst_b, _ = nodes.create.with_transform.locator(name="dst_b", parent=dst_p)
-
-    for node in [src_p, src, dst_p]:
-        node.translate.set(1, 2, 3)
-        node.translate.set(
-            random.random(),
-            random.random(),
-            random.random(),
-        )
-        node.rotate.set(
-            random.randint(-360, 360),
-            random.randint(-360, 360),
-            random.randint(-360, 360),
-        )
-        node.scale.set(
-            random.random(),
-            random.random(),
-            random.random(),
-        )
+    src.t.connect(dst.t)
+    src.r.connect(dst.r)
+    src.s.connect(dst.s)
+    src.sh.connect(dst.sh)
 
     nodes.modifier_manager.do_it_dag()
     nodes.modifier_manager.do_it_dg()
 
-    # dst_a
-    src_wm = src.wm[0].get()
-    dst_pim = dst_a.pim[0].get()
-    m = src_wm * dst_pim
-    dst_a.translate.set(m.translate)
-    dst_a.rotate.set(m.rotate)
-    dst_a.scale.set(m.scale)
-    dst_a.shear.set(m.shear)
+    logger.debug("----")
+    src_plug = dst.t.src_plug()
+    logger.debug(f"src_plug      : {src_plug}")
+    logger.debug(f"type(src_plug): {type(src_plug)}")
 
-    # dst_b
-    m = src.get_local_matrix(dst_b)
-    dst_b.translate.set(m.translate)
-    dst_b.rotate.set(m.rotate)
-    dst_b.scale.set(m.scale)
-    dst_b.shear.set(m.shear)
+    src_name = dst.t.src_name()
+    logger.debug(f"src_name      : {src_name}")
+    logger.debug(f"type(src_name): {type(src_name)}")
 
-    nodes.modifier_manager.do_it_dg()
+    src_plug_name = dst.t.src_plug_name()
+    logger.debug(f"src_plug_name      : {src_plug_name}")
+    logger.debug(f"type(src_plug_name): {type(src_plug_name)}")
+
+    logger.debug("----")
+    src_plug = dst.t.src_plug(filter_type=nodes.types.Joint)
+    logger.debug(f"src_plug      : {src_plug}")
+    logger.debug(f"type(src_plug): {type(src_plug)}")
+
+    src_name = dst.t.src_name(filter_type=nodes.types.Joint)
+    logger.debug(f"src_name      : {src_name}")
+    logger.debug(f"type(src_name): {type(src_name)}")
+
+    src_plug_name = dst.t.src_plug_name(filter_type=nodes.types.Joint)
+    logger.debug(f"src_plug_name      : {src_plug_name}")
+    logger.debug(f"type(src_plug_name): {type(src_plug_name)}")
+
+    logger.debug("----")
+    dst_plugs = src.t.dst_plugs()
+    logger.debug(f"dst_plugs      : {dst_plugs}")
+    logger.debug(f"type(dst_plugs): {type(dst_plugs)}")
+
+    dst_names = src.t.dst_names()
+    logger.debug(f"dst_names      : {dst_names}")
+    logger.debug(f"type(dst_names): {type(dst_names)}")
+
+    dst_plug_names = dst.t.dst_plug_names()
+    logger.debug(f"dst_plug_names      : {dst_plug_names}")
+    logger.debug(f"type(dst_plug_names): {type(dst_plug_names)}")
+
+    logger.debug("----")
+    dst_plugs = src.t.dst_plugs(
+        filter_type=nodes.types.Transform, include_subclasses=False
+    )
+    logger.debug(f"dst_plugs      : {dst_plugs}")
+    logger.debug(f"type(dst_plugs): {type(dst_plugs)}")
+
+    dst_names = src.t.dst_names(
+        filter_type=nodes.types.Transform, include_subclasses=False
+    )
+    logger.debug(f"dst_names      : {dst_names}")
+    logger.debug(f"type(dst_names): {type(dst_names)}")
+
+    dst_plug_names = dst.t.dst_plug_names(
+        filter_type=nodes.types.Transform, include_subclasses=False
+    )
+    logger.debug(f"dst_plug_names      : {dst_plug_names}")
+    logger.debug(f"type(dst_plug_names): {type(dst_plug_names)}")